@@ -1,6 +1,6 @@
 {
   "name": "ngx-maplibre-gl-srcs",
-  "version": "13.0.0",
+  "version": "14.0.0-pre.1",
   "scripts": {
     "ng": "ng",
     "start": "ng serve showcase",
@@ -29,49 +29,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.2",
-    "@angular/cdk": "^13.1.2",
-    "@angular/common": "^13.1.2",
-    "@angular/compiler": "^13.1.2",
-    "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.1.2",
-    "@angular/material": "^13.1.2",
-    "@angular/platform-browser": "^13.1.2",
-    "@angular/platform-browser-dynamic": "^13.1.2",
-    "@angular/router": "^13.1.2",
-    "@ngrx/effects": "^13.0.2",
-    "@ngrx/router-store": "^13.0.2",
-    "@ngrx/store": "^13.0.2",
-    "@ngrx/store-devtools": "^13.0.2",
+    "@angular/animations": "^14.2.5",
+    "@angular/cdk": "^14.2.4",
+    "@angular/common": "^14.2.5",
+    "@angular/compiler": "^14.2.5",
+    "@angular/core": "^14.2.5",
+    "@angular/forms": "^14.2.5",
+    "@angular/material": "^14.2.4",
+    "@angular/platform-browser": "^14.2.5",
+    "@angular/platform-browser-dynamic": "^14.2.5",
+    "@angular/router": "^14.2.5",
+    "@ngrx/effects": "^14.3.2",
+    "@ngrx/router-store": "^14.3.2",
+    "@ngrx/store": "^14.3.2",
+    "@ngrx/store-devtools": "^14.3.2",
     "@stackblitz/sdk": "^1.6.0",
     "@types/geojson": "^7946.0.7",
     "lodash-es": "^4.17.11",
-    "maplibre-gl": "^2.1.9",
-    "ngx-markdown": "^13.0.0",
-    "rxjs": "^7.4.0",
+    "maplibre-gl": "^3.0.0-pre.1",
+    "ngx-markdown": "^14.0.1",
+    "rxjs": "7.5.7",
     "scroll-into-view-if-needed": "^2.2.20",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.3",
-    "@angular-eslint/builder": "13.0.1",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/schematics": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "^13.1.3",
-    "@angular/compiler-cli": "^13.1.2",
-    "@ngrx/schematics": "^13.0.2",
+    "@angular-devkit/build-angular": "^14.2.5",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.5",
+    "@angular/compiler-cli": "^14.2.5",
+    "@ngrx/schematics": "^14.3.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.6",
     "@types/lodash-es": "^4.17.3",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
     "angular-cli-ghpages": "^0.6.0",
     "commitizen": "^4.0.3",
     "cz-conventional-changelog": "^3.1.0",
-    "eslint": "^8.2.0",
+    "eslint": "^8.23.0",
     "husky": "^4.2.1",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -82,14 +82,14 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "karma-mocha-reporter": "^2.2.5",
     "lint-staged": "^10.0.7",
-    "ng-packagr": "^13.1.3",
+    "ng-packagr": "^14.2.1",
     "pixel-diff": "^1.0.1",
     "prettier": "2.2.1",
     "protractor": "~7.0.0",
     "protractor-browser-logs": "^1.0.446",
     "standard-version": "^9.1.0",
     "ts-node": "^9.1.1",
-    "typescript": "~4.5.4",
+    "typescript": "~4.8.4",
     "typescript-tslint-plugin": "^1.0.0"
   },
   "config": {
